+++ tket/src/Transformations/include/Transformations/BasicOptimisation.hpp
@@ -137,6 +137,12 @@ Transform ZZPhase_to_Rz();
  */
 Transform normalise_TK2();
 
+/**
+ * @brief Squash single qubit gates into PhasedX and Rz gates.
+ * Commute Rzs to the back if possible.
+ */
+Transform squash_1qb_to_Rz_PhasedX();
+
 }  // namespace Transforms
 
 }  // namespace tket